+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+
+## [1.2.4] - 2021-09-29
+- Pull image fixes from FLTK.
+- Add app::compose, compose_state, reset_marked_text and insertion_point_location.
+- Add app::copy.
+- Check the size of drawn text.
+- Get scaling factor on MacOS without requiring info.plist.
+- Add copy and paste overloads for selection buffers.
+- Add app::add_clipboard_notify() and remove_clipboard_notify().
+
 ## [1.2.3] - 2021-09-24
 - Add a new group::Flex widget wrapping [Fl_Flex](https://github.com/osen/FL_Flex).
 - Use fork of FL_Flex to avoid patching, then linking libstd++.